+++ ionoscloud/provider.go
@@ -19,6 +19,7 @@ import (
 	dataplatformService "github.com/ionos-cloud/terraform-provider-ionoscloud/v6/services/dataplatform"
 	dbaasService "github.com/ionos-cloud/terraform-provider-ionoscloud/v6/services/dbaas"
 	dnsService "github.com/ionos-cloud/terraform-provider-ionoscloud/v6/services/dns"
+	loggingService "github.com/ionos-cloud/terraform-provider-ionoscloud/v6/services/logging"
 	"github.com/ionos-cloud/terraform-provider-ionoscloud/v6/utils"
 )
 
@@ -32,6 +33,7 @@ type SdkBundle struct {
 	ContainerClient    *crService.Client
 	DataplatformClient *dataplatformService.Client
 	DNSClient          *dnsService.Client
+	LoggingClient      *loggingService.Client
 }
 
 type ClientOptions struct {
@@ -103,6 +105,8 @@ func Provider() *schema.Provider {
 			NetworkLoadBalancerResource: resourceNetworkLoadBalancer(),
 			NetworkLoadBalancerForwardingRuleResource: resourceNetworkLoadBalancerForwardingRule(),
 			PsqlClusterResource:                       resourceDbaasPgSqlCluster(),
+			PsqlUserResource:                          resourceDbaasPgSqlUser(),
+			PsqlDatabaseResource:                      resourceDbaasPgSqlDatabase(),
 			DBaasMongoClusterResource:                 resourceDbaasMongoDBCluster(),
 			DBaasMongoUserResource:                    resourceDbaasMongoUser(),
 			ALBResource:                               resourceApplicationLoadBalancer(),
@@ -115,6 +119,7 @@ func Provider() *schema.Provider {
 			DataplatformNodePoolResource:              resourceDataplatformNodePool(),
 			DNSZoneResource:                           resourceDNSZone(),
 			DNSRecordResource:                         resourceDNSRecord(),
+			LoggingPipelineResource:                   resourceLoggingPipeline(),
 		},
 		DataSourcesMap: map[string]*schema.Resource{
 			DatacenterResource:                        dataSourceDataCenter(),
@@ -146,6 +151,9 @@ func Provider() *schema.Provider {
 			ShareResource:                             dataSourceShare(),
 			ResourceIpFailover:                        dataSourceIpFailover(),
 			PsqlClusterResource:                       dataSourceDbaasPgSqlCluster(),
+			PsqlUserResource:                          dataSourceDbaasPgSqlUser(),
+			PsqlDatabaseResource:                      dataSourceDbaasPgSqlDatabase(),
+			PsqlDatabasesResource:                     dataSourceDbaasPgSqlDatabases(),
 			DBaasMongoClusterResource:                 dataSourceDbaasMongoCluster(),
 			DBaaSMongoTemplateResource:                dataSourceDbassMongoTemplate(),
 			PsqlVersionsResource:                      dataSourceDbaasPgSqlVersions(),
@@ -164,6 +172,7 @@ func Provider() *schema.Provider {
 			DataplatformVersionsDataSource:            dataSourceDataplatformVersions(),
 			DNSZoneDataSource:                         dataSourceDNSZone(),
 			DNSRecordDataSource:                       dataSourceDNSRecord(),
+			LoggingPipelineDataSource:                 dataSourceLoggingPipeline(),
 		},
 	}
 
@@ -222,6 +231,7 @@ func providerConfigure(d *schema.ResourceData, terraformVersion string) (interfa
 		ContainerClient:    NewClientByType(clientOpts, containerRegistryClient).(*crService.Client),
 		DataplatformClient: NewClientByType(clientOpts, dataplatformClient).(*dataplatformService.Client),
 		DNSClient:          NewClientByType(clientOpts, dnsClient).(*dnsService.Client),
+		LoggingClient:      NewClientByType(clientOpts, loggingClient).(*loggingService.Client),
 	}, nil
 }
 
@@ -253,6 +263,8 @@ func NewClientByType(clientOpts ClientOptions, clientType clientType) interface{
 		return dataplatformService.NewClient(clientOpts.Username, clientOpts.Password, clientOpts.Token, clientOpts.Url, clientOpts.Version, clientOpts.TerraformVersion)
 	case dnsClient:
 		return dnsService.NewClient(clientOpts.Username, clientOpts.Password, clientOpts.Token, clientOpts.Url, clientOpts.Version, clientOpts.TerraformVersion)
+	case loggingClient:
+		return loggingService.NewClient(clientOpts.Username, clientOpts.Password, clientOpts.Token, clientOpts.Url, clientOpts.Version, clientOpts.TerraformVersion)
 	default:
 		log.Fatalf("[ERROR] unknown client type %d", clientType)
 	}