+++ CHANGELOG.md
@@ -1,6 +1,23 @@
-## 6.4.3
+## 6.4.5
+### Features
 - Make `nic` list in `ionoscloud_server` resource optional
 - Make `firewall` list in `ionoscloud_server` resource optional and allow multiple inline firewall rules in the list
+
+## 6.4.4
+
+### Features
+- Add support for PgSQL User & Databases
+
+### Dependency update
+- Update `sdk-go-dbaas-postgres` to version 1.1.2
+
+## 6.4.3
+
+### Documentation
+- Improve example for `ionoscloud_private_crossconnect`
+
+### Fixes
+- Remove unpopulated `credentials` field from mongodb cluster data source.
 
 ## 6.4.2
 ### Fixes